move TextureAtlasPlugin to texture_atlas.rs
